Clean master branch
So far, we have done the first phase of our project. We wrote python programs to fetch the data from glassdoor.com and created a database of business analyst jobs in 30 major cities in the US. We are going to use the file BA_job_clean.csv as the dataset to move into our second part, doing some fundamental data analysis.